+++ src/tools/general.js
@@ -6,16 +6,11 @@ const check = require('check-types');
 
 const {
     callForReview,
-    getPlacePrices,
     buildHotelUrl,
     buildRestaurantUrl,
-    getPlaceInformation,
     buildAttractionsUrl,
-    callForAttractionList,
-    callForAttractionReview,
     getAgentOptions,
     getReviewTagsForLocation,
-    callForRestaurantList,
 } = require('./api');
 
 const { utils: { log } } = Apify;
@@ -63,17 +58,6 @@ async function resolveInBatches(promiseArray, batchLength = 10) {
         if (promises.length % batchLength === 0) await Promise.all(promises);
     }
     return Promise.all(promises);
-}
-
-function getHotelIds($) {
-    const hotelIds = [];
-    const divs = $('.prw_rup.prw_meta_hsx_responsive_listing.ui_section.listItem');
-    divs.each((index, element) => {
-        const el = $(element).find('.meta_listing');
-        const placeId = el.attr('data-locationid');
-        hotelIds.push(placeId);
-    });
-    return hotelIds;
 }
 
 const processReview = (review, remoteId) => {
@@ -178,68 +162,6 @@ async function getReviews(id, client) {
     return result;
 }
 
-async function processHotel(placeInfo, client, dataset) {
-    const { location_id: id } = placeInfo;
-    let reviews = [];
-    let placePrices;
-
-    try {
-        placePrices = await getPlacePrices(id, randomDelay);
-    } catch (e) {
-        log.warning('Hotels: Could not get place prices', { errorMessage: e.message });
-    }
-
-    if (global.INCLUDE_REVIEWS) {
-        try {
-            reviews = await getReviews(id, client);
-        } catch (e) {
-            log.error('Could not get reviews', e);
-        }
-    }
-
-    if (!placeInfo) {
-        return;
-    }
-    const prices = placePrices ? placePrices.offers.map(offer => ({
-        provider: offer.provider_display_name,
-        price: offer.display_price_int ? offer.display_price_int : 'NOT_PROVIDED',
-        isBookable: offer.is_bookable,
-        link: offer.link,
-    })) : [];
-    const place = {
-        id: placeInfo.location_id,
-        type: 'HOTEL',
-        name: placeInfo.name,
-        awards: placeInfo.awards && placeInfo.awards.map(award => ({ year: award.year, name: award.display_name })),
-        rankingPosition: placeInfo.ranking_position,
-        priceLevel: placeInfo.price_level,
-        category: placeInfo.ranking_category,
-        rating: placeInfo.rating,
-        hotelClass: placeInfo.hotel_class,
-        hotelClassAttribution: placeInfo.hotel_class_attribution,
-        phone: placeInfo.phone,
-        address: placeInfo.address,
-        email: placeInfo.email,
-        amenities: placeInfo.amenities && placeInfo.amenities.map(amenity => amenity.name),
-        prices,
-        latitude: placeInfo.latitude,
-        longitude: placeInfo.longitude,
-        webUrl: placeInfo.web_url,
-        website: placeInfo.website,
-        reviews,
-    };
-    if (global.INCLUDE_REVIEW_TAGS) {
-        const tags = await getReviewTags(id);
-        place.reviewTags = tags;
-    }
-    log.debug('Data for hotel: ', place);
-    if (dataset) {
-        await dataset.pushData(place);
-    } else {
-        await Apify.setValue('OUTPUT', JSON.stringify(place), { contentType: 'application/json' });
-    }
-}
-
 
 function getRequestListSources(locationId, includeHotels, includeRestaurants, includeAttractions) {
     const sources = [];
@@ -264,84 +186,6 @@ function getRequestListSources(locationId, includeHotels, includeRestaurants, in
         });
     }
     return sources;
-}
-
-function getRestaurantIds($) {
-    const ids = [];
-    $('.listing.rebrand').each((index, element) => {
-        const id = $(element).attr('id');
-        if (id) {
-            const split = id.split('_');
-            ids.push(split[split.length - 1]);
-        }
-    });
-    return ids;
-}
-
-function getHours(placeInfo) {
-    const placeHolder = [];
-
-    if (!placeInfo.hours) {
-        return placeHolder;
-    }
-
-    if (!placeInfo.hours.week_ranges) {
-        return placeHolder;
-    }
-
-    return placeInfo.hours.week_ranges.map(wR => wR.map(day => ({ open: day.open_time, close: day.close_time })));
-}
-
-
-async function processRestaurant(placeInfo, client, dataset) {
-    const { location_id: id } = placeInfo;
-    let reviews = [];
-    if (global.INCLUDE_REVIEWS) {
-        try {
-            reviews = await getReviews(id, client);
-        } catch (e) {
-            log.error('Could not get reviews', e);
-        }
-    }
-    if (!placeInfo) {
-        return;
-    }
-    const place = {
-        id: placeInfo.location_id,
-        type: 'RESTAURANT',
-        name: placeInfo.name,
-        awards: placeInfo.awards && placeInfo.awards.map(award => ({ year: award.year, name: award.display_name })),
-        rankingPosition: placeInfo.ranking_position,
-        priceLevel: placeInfo.price_level,
-        category: placeInfo.ranking_category,
-        rating: placeInfo.rating,
-        isClosed: placeInfo.is_closed,
-        isLongClosed: placeInfo.is_long_closed,
-        phone: placeInfo.phone,
-        address: placeInfo.address,
-        email: placeInfo.email,
-        cuisine: placeInfo.cuisine && placeInfo.cuisine.map(cuisine => cuisine.name),
-        mealTypes: placeInfo.mealTypes && placeInfo.mealTypes.map(m => m.name),
-        hours: getHours(placeInfo),
-        latitude: placeInfo.latitude,
-        longitude: placeInfo.longitude,
-        webUrl: placeInfo.web_url,
-        website: placeInfo.website,
-        numberOfReviews: placeInfo.num_reviews,
-        rankingDenominator: placeInfo.ranking_denominator,
-        reviews,
-    };
-    if (global.INCLUDE_REVIEW_TAGS) {
-        const tags = await getReviewTags(id);
-        place.reviewTags = tags;
-    }
-    log.debug('Data for restaurant: ', place);
-
-    if (dataset) {
-        await dataset.pushData(place);
-    } else {
-        await Apify.setValue('OUTPUT', JSON.stringify(place), { contentType: 'application/json' });
-    }
 }
 
 async function getClient() {
@@ -417,25 +261,6 @@ function validateInput(input) {
     log.info('Input validation OK');
 }
 
-async function getAttractions(locationId) {
-    let attractions = [];
-    let offset = 0;
-    const limit = 20;
-    const data = await callForAttractionList(locationId, limit);
-    attractions = attractions.concat(data.data);
-    if (data.paging && data.paging.next) {
-        const totalResults = data.paging.total_results;
-        const numberOfRuns = Math.ceil(totalResults / limit);
-        log.info(`Going to process ${numberOfRuns} pages of attractions`);
-        for (let i = 0; i <= numberOfRuns; i++) {
-            offset += limit;
-            const data2 = await callForAttractionList(locationId, limit, offset);
-            attractions = attractions.concat(data2.data);
-        }
-    }
-    return attractions;
-}
-
 async function getReviewTags(locationId) {
     let tags = [];
     let offset = 0;
@@ -454,205 +279,14 @@ async function getReviewTags(locationId) {
     }
     return tags;
 }
-function processAttractionReview(review) {
-    const {
-        lang,
-        text,
-        published_date: publishedDate,
-        rating,
-        travel_date: travelDate,
-        user,
-        title,
-        machine_translated: machineTranslated,
-        subratings,
-    } = review;
-
-    return {
-        language: lang,
-        title,
-        text,
-        publishedDate,
-        rating,
-        travelDate,
-        user: {
-            username: user.username,
-            helpfulVotes: user.helpful_votes,
-
-        },
-        subratings,
-        machineTranslated,
-    };
-}
-
-async function getReviewsForAttraction(locationId) {
-    const reviews = [];
-    let offset = 0;
-    const limit = 50;
-    const data = await callForAttractionReview(locationId, limit);
-    let { data: revs } = data;
-    let lastIndex = findLastReviewIndex(revs, 'published_date');
-    let shouldSlice = lastIndex >= 0;
-    if (shouldSlice) {
-        revs = revs.slice(0, lastIndex);
-    }
-    revs.forEach(review => reviews.push(processAttractionReview(review)));
-    if (shouldSlice) return reviews;
-    if (data.paging && data.paging.next) {
-        const totalResults = data.paging.total_results;
-        const numberOfRuns = Math.ceil(totalResults / limit);
-        log.info(`Going to process ${numberOfRuns} pages of reviews`);
-        for (let i = 0; i <= numberOfRuns; i++) {
-            offset += limit;
-            let { data: reviewsToPush } = await callForAttractionReview(locationId, limit, offset);
-            lastIndex = findLastReviewIndex(reviewsToPush, 'published_date');
-            shouldSlice = lastIndex >= 0;
-            if (shouldSlice) {
-                reviewsToPush = reviewsToPush.slice(0, lastIndex);
-            }
-            reviewsToPush.forEach(review => reviews.push(processAttractionReview(review)));
-            if (shouldSlice) break;
-        }
-    }
-    return reviews;
-}
-
-async function getAttractionDetail(attraction) {
-    log.info(`Processing detail for ${attraction.name} attraction`);
-    const locationId = attraction.location_id;
-    let reviews = [];
-    if (global.INCLUDE_REVIEWS) {
-        try {
-            reviews = await getReviewsForAttraction(locationId);
-            log.info(`Got ${reviews.length} reviews for ${attraction.name}`);
-        } catch (e) {
-            log.error(`Could not get reviews for attraction ${attraction.name} due to ${e.message}`);
-        }
-    }
-
-    attraction.reviews = reviews;
-    return attraction;
-}
-
-async function processAttraction(attraction) {
-    const attr = await getAttractionDetail(attraction);
-    if (global.includeTagsInReviews) {
-    }
-    return Apify.pushData(attr);
-}
-async function getAttractions(locationId) {
-    let attractions = [];
-    let offset = 0;
-    const limit = 20;
-    const data = await callForAttractionList(locationId, limit);
-    attractions = attractions.concat(data.data);
-    console.log(data.paging, 'PAGING');
-    if (data.paging && data.paging.next) {
-        const totalResults = data.paging.total_results;
-        const numberOfRuns = Math.ceil(totalResults / limit);
-        log.info(`Going to process ${numberOfRuns} pages of attractions`);
-        for (let i = 0; i <= numberOfRuns; i++) {
-            offset += limit;
-            const data2 = await callForAttractionList(locationId, limit, offset);
-            attractions = attractions.concat(data2.data);
-        }
-    }
-    return attractions;
-}
-
-function processAttractionReview(review) {
-    const {
-        lang,
-        text,
-        published_date: publishedDate,
-        rating,
-        travel_date: travelDate,
-        user,
-        title,
-        machine_translated: machineTranslated,
-        subratings,
-    } = review;
-
-    return {
-        language: lang,
-        title,
-        text,
-        publishedDate,
-        rating,
-        travelDate,
-        user: {
-            username: user.username,
-            helpfulVotes: user.helpful_votes,
-
-        },
-        subratings,
-        machineTranslated,
-    };
-}
-
-async function getReviewsForAttraction(locationId) {
-    const reviews = [];
-    let offset = 0;
-    const limit = 50;
-    const data = await callForAttractionReview(locationId, limit);
-    let { data: revs } = data;
-    let lastIndex = findLastReviewIndex(revs, 'published_date');
-    let shouldSlice = lastIndex >= 0;
-    if (shouldSlice) {
-        revs = revs.slice(0, lastIndex);
-    }
-    revs.forEach(review => reviews.push(processAttractionReview(review)));
-    if (shouldSlice) return reviews;
-    if (data.paging && data.paging.next) {
-        const totalResults = data.paging.total_results;
-        const numberOfRuns = Math.ceil(totalResults / limit);
-        log.info(`Going to process ${numberOfRuns} pages of reviews`);
-        for (let i = 0; i <= numberOfRuns; i++) {
-            offset += limit;
-            let { data: reviewsToPush } = await callForAttractionReview(locationId, limit, offset);
-            lastIndex = findLastReviewIndex(reviewsToPush, 'published_date');
-            shouldSlice = lastIndex >= 0;
-            if (shouldSlice) {
-                reviewsToPush = reviewsToPush.slice(0, lastIndex);
-            }
-            reviewsToPush.forEach(review => reviews.push(processAttractionReview(review)));
-            if (shouldSlice) break;
-        }
-    }
-    return reviews;
-}
-
-async function getAttractionDetail(attraction) {
-    log.info(`Processing detail for ${attraction.name} attraction`);
-    const locationId = attraction.location_id;
-    let reviews = [];
-    if (global.INCLUDE_REVIEWS) {
-        try {
-            reviews = await getReviewsForAttraction(locationId);
-            log.info(`Got ${reviews.length} reviews for ${attraction.name}`);
-        } catch (e) {
-            log.error(`Could not get reviews for attraction ${attraction.name} due to ${e.message}`);
-        }
-    }
-
-    attraction.reviews = reviews;
-    return attraction;
-}
-
-async function processAttraction(attraction) {
-    const attr = await getAttractionDetail(attraction);
-    return Apify.pushData(attr);
-}
 
 module.exports = {
     resolveInBatches,
-    getHotelIds,
-    processHotel,
     getRequestListSources,
-    getRestaurantIds,
-    processRestaurant,
     getClient,
     randomDelay,
     validateInput,
-    getAttractions,
-    processAttraction,
+    getReviewTags,
+    getReviews,
+    findLastReviewIndex,
 };
